Listar

+++ src/main/java/dao/EmpresaJpaController.java
@@ -4,38 +4,65 @@
  */
 package dao;
 
+import dao.exceptions.IllegalOrphanException;
 import dao.exceptions.NonexistentEntityException;
-import dto.Empresa;
 import java.io.Serializable;
-import java.util.List;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import dto.Empleado;
+import dto.Empresa;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
+import javax.persistence.Persistence;
 
 /**
  *
- * @author redcr
+ * @author USER
  */
 public class EmpresaJpaController implements Serializable {
 
     public EmpresaJpaController(EntityManagerFactory emf) {
         this.emf = emf;
     }
-    private EntityManagerFactory emf = null;
+
+    public EmpresaJpaController() {
+    }
+    
+    private EntityManagerFactory emf = Persistence.createEntityManagerFactory("com.mycompany_Asistencia03_war_1.0-SNAPSHOTPU");
 
     public EntityManager getEntityManager() {
         return emf.createEntityManager();
     }
 
     public void create(Empresa empresa) {
+        if (empresa.getEmpleadoList() == null) {
+            empresa.setEmpleadoList(new ArrayList<Empleado>());
+        }
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            List<Empleado> attachedEmpleadoList = new ArrayList<Empleado>();
+            for (Empleado empleadoListEmpleadoToAttach : empresa.getEmpleadoList()) {
+                empleadoListEmpleadoToAttach = em.getReference(empleadoListEmpleadoToAttach.getClass(), empleadoListEmpleadoToAttach.getIdEmpleado());
+                attachedEmpleadoList.add(empleadoListEmpleadoToAttach);
+            }
+            empresa.setEmpleadoList(attachedEmpleadoList);
             em.persist(empresa);
+            for (Empleado empleadoListEmpleado : empresa.getEmpleadoList()) {
+                Empresa oldIdempresaOfEmpleadoListEmpleado = empleadoListEmpleado.getIdempresa();
+                empleadoListEmpleado.setIdempresa(empresa);
+                empleadoListEmpleado = em.merge(empleadoListEmpleado);
+                if (oldIdempresaOfEmpleadoListEmpleado != null) {
+                    oldIdempresaOfEmpleadoListEmpleado.getEmpleadoList().remove(empleadoListEmpleado);
+                    oldIdempresaOfEmpleadoListEmpleado = em.merge(oldIdempresaOfEmpleadoListEmpleado);
+                }
+            }
             em.getTransaction().commit();
         } finally {
             if (em != null) {
@@ -44,17 +71,50 @@ public class EmpresaJpaController implements Serializable {
         }
     }
 
-    public void edit(Empresa empresa) throws NonexistentEntityException, Exception {
+    public void edit(Empresa empresa) throws IllegalOrphanException, NonexistentEntityException, Exception {
         EntityManager em = null;
         try {
             em = getEntityManager();
             em.getTransaction().begin();
+            Empresa persistentEmpresa = em.find(Empresa.class, empresa.getIdempresa());
+            List<Empleado> empleadoListOld = persistentEmpresa.getEmpleadoList();
+            List<Empleado> empleadoListNew = empresa.getEmpleadoList();
+            List<String> illegalOrphanMessages = null;
+            for (Empleado empleadoListOldEmpleado : empleadoListOld) {
+                if (!empleadoListNew.contains(empleadoListOldEmpleado)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Empleado " + empleadoListOldEmpleado + " since its idempresa field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            List<Empleado> attachedEmpleadoListNew = new ArrayList<Empleado>();
+            for (Empleado empleadoListNewEmpleadoToAttach : empleadoListNew) {
+                empleadoListNewEmpleadoToAttach = em.getReference(empleadoListNewEmpleadoToAttach.getClass(), empleadoListNewEmpleadoToAttach.getIdEmpleado());
+                attachedEmpleadoListNew.add(empleadoListNewEmpleadoToAttach);
+            }
+            empleadoListNew = attachedEmpleadoListNew;
+            empresa.setEmpleadoList(empleadoListNew);
             empresa = em.merge(empresa);
+            for (Empleado empleadoListNewEmpleado : empleadoListNew) {
+                if (!empleadoListOld.contains(empleadoListNewEmpleado)) {
+                    Empresa oldIdempresaOfEmpleadoListNewEmpleado = empleadoListNewEmpleado.getIdempresa();
+                    empleadoListNewEmpleado.setIdempresa(empresa);
+                    empleadoListNewEmpleado = em.merge(empleadoListNewEmpleado);
+                    if (oldIdempresaOfEmpleadoListNewEmpleado != null && !oldIdempresaOfEmpleadoListNewEmpleado.equals(empresa)) {
+                        oldIdempresaOfEmpleadoListNewEmpleado.getEmpleadoList().remove(empleadoListNewEmpleado);
+                        oldIdempresaOfEmpleadoListNewEmpleado = em.merge(oldIdempresaOfEmpleadoListNewEmpleado);
+                    }
+                }
+            }
             em.getTransaction().commit();
         } catch (Exception ex) {
             String msg = ex.getLocalizedMessage();
             if (msg == null || msg.length() == 0) {
-                Integer id = empresa.getIdEmpresa();
+                Integer id = empresa.getIdempresa();
                 if (findEmpresa(id) == null) {
                     throw new NonexistentEntityException("The empresa with id " + id + " no longer exists.");
                 }
@@ -67,7 +127,7 @@ public class EmpresaJpaController implements Serializable {
         }
     }
 
-    public void destroy(Integer id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws IllegalOrphanException, NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
@@ -75,9 +135,20 @@ public class EmpresaJpaController implements Serializable {
             Empresa empresa;
             try {
                 empresa = em.getReference(Empresa.class, id);
-                empresa.getIdEmpresa();
+                empresa.getIdempresa();
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The empresa with id " + id + " no longer exists.", enfe);
+            }
+            List<String> illegalOrphanMessages = null;
+            List<Empleado> empleadoListOrphanCheck = empresa.getEmpleadoList();
+            for (Empleado empleadoListOrphanCheckEmpleado : empleadoListOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Empresa (" + empresa + ") cannot be destroyed since the Empleado " + empleadoListOrphanCheckEmpleado + " in its empleadoList field has a non-nullable idempresa field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(empresa);
             em.getTransaction().commit();
@@ -133,5 +204,22 @@ public class EmpresaJpaController implements Serializable {
             em.close();
         }
     }
-    
+      
+    public List<Object[]> listar() {
+        EntityManager em = getEntityManager();
+        try {
+            Query q = em.createNamedQuery("Empresa.listar");
+            return q.getResultList();
+        } finally {
+            em.close();
+        }
+    }
+    public static void main(String[] args) {
+        EmpresaJpaController Empresa=new EmpresaJpaController();
+        List<Object[]> listar =Empresa.listar();
+        for(Object[] object:listar){
+            System.out.println(Arrays.toString(object));
+        }
+    }
+
 }
